feat: work on alloysmelter, fix energy, more

+++ astromine-core/src/main/java/com/github/mixinors/astromine/common/block/entity/machine/EnergySizeProvider.java
@@ -25,5 +25,5 @@
 package com.github.mixinors.astromine.common.block.entity.machine;
 
 public interface EnergySizeProvider {
-	double getEnergySize();
+	long getEnergySize();
 }
